# main-webbase/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        example: invalid body
        type: string
    type: object
  dto.EventCreateResult:
    properties:
      event:
        $ref: '#/definitions/models.Event'
      form_id:
        type: string
      organizer_count:
        type: integer
      schedules:
        items:
          $ref: '#/definitions/models.EventSchedule'
        type: array
    type: object
  dto.EventDetail:
    properties:
      current_participation:
        type: integer
      description:
        type: string
      event_id:
        type: string
      have_form:
        type: boolean
      max_participation:
        type: integer
      orgpath:
        type: string
      picture_url:
        type: string
      posted_as:
        $ref: '#/definitions/models.PostedAs'
      schedules:
        items:
          $ref: '#/definitions/models.EventSchedule'
        type: array
      status:
        type: string
      topic:
        type: string
      user_participants:
        items:
          $ref: '#/definitions/dto.UserParticipant'
        type: array
      visibility:
        $ref: '#/definitions/models.Visibility'
    type: object
  dto.EventRequestDTO:
    properties:
      description:
        example: A workshop on AI applications
        type: string
      have_form:
        example: true
        type: boolean
      max_participation:
        example: 50
        type: integer
      node_id:
        example: 66ffa43e9a7c39b1d87f6401
        type: string
      org_of_content:
        example: /fac/eng/com
        type: string
      picture_url:
        example: http://45.144.166.252:46602/uploads/cat.png
        type: string
      posted_as:
        $ref: '#/definitions/models.PostedAs'
      schedules:
        items:
          properties:
            date:
              example: "2025-10-15T00:00:00Z"
              type: string
            description:
              example: Morning session
              type: string
            location:
              example: Innovation Building Room 301
              type: string
            time_end:
              example: "2025-10-15T12:00:00Z"
              type: string
            time_start:
              example: "2025-10-15T09:00:00Z"
              type: string
          type: object
        type: array
      status:
        enum:
        - active
        - draft
        - inactive
        example: draft
        type: string
      topic:
        example: AI Workshop
        type: string
      visibility:
        $ref: '#/definitions/models.Visibility'
    required:
    - node_id
    - topic
    type: object
...
  dto.UserParticipant:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      status:
        description: '"participant/organizer"'
        type: string
      user_id:
        type: string
    type: object
...
  models.AudienceItem:
    properties:
      org_path:
        type: string
      scope:
        description: '"exact" | "subtree"'
        type: string
    type: object
...
  models.Event:
    properties:
      created_at:
        type: string
      description:
        type: string
      have_form:
        type: boolean
      id:
        type: string
      max_participation:
        type: integer
      node_id:
        type: string
      org_of_content:
        type: string
      picture_url:
        type: string
      posted_as:
        $ref: '#/definitions/models.PostedAs'
      status:
        description: active, draft, inactive
        type: string
      topic:
        type: string
      updated_at:
        type: string
      visibility:
        $ref: '#/definitions/models.Visibility'
    type: object
...
  models.EventSchedule:
    properties:
      date:
        type: string
      description:
        type: string
      event_id:
        type: string
      id:
        type: string
      location:
        type: string
      time_end:
        type: string
      time_start:
        type: string
    type: object
...
  models.PostedAs:
    properties:
      label:
        type: string
      org_path:
        type: string
      position_key:
        type: string
    type: object
...
  models.Visibility:
    properties:
      access:
        type: string
      allow_user_ids:
        items:
          type: string
        type: array
      audience:
        items:
          $ref: '#/definitions/models.AudienceItem'
        type: array
      deny_user_ids:
        items:
          type: string
        type: array
      exclude_positions:
        items:
          type: string
        type: array
      include_positions:
        items:
          type: string
        type: array
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is a sample server for user management.
  title: Fiber MongoDB API
  version: "1.0"
paths:
...
  /event:
    get:
      consumes:
      - application/json
      description: 'Retrieve all events that the current user can see. This endpoint
        returns events visible to the current authenticated user. You can optionally
        filter the events using query parameters: \n - `q` (string): Search text that
        matches the event''s topic or description (case-insensitive). \n - `role`
        (string, comma-separated): Filter by user role or organization path. Each
        value can be: \n - A position key, e.g., `Lecturer` (matches `postedas.position_key`,
        case-insensitive exact match). \n - An organization path, e.g., `/fac/eng/com`
        (matches `org_of_content`). \n - Subtree prefix is supported using `/*`, e.g.,
        `/fac/eng/*` matches `/fac/eng/com` or `/fac/eng/math`. \n If no filters are
        applied, all events that the user can see (based on visibility rules) are
        returned. \n Visibility rules: \n - `public`: visible to everyone. \n - `org`:
        visible only to users whose organization is included in the audience. \n -
        `draft`: visible only to organizers within the same organization.'
      parameters:
      - description: Search text in topic or description
        in: query
        name: q
        type: string
      - description: Comma-separated list of roles or org paths to filter
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Array of visible events with their next upcoming schedule
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized - user not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all visible events
      tags:
      - events
    post:
      consumes:
      - multipart/form-data
      description: Create an event with optional image upload and schedule creation
      parameters:
      - description: Event image file upload
        in: formData
        name: file
        type: file
      - description: Node ID
        in: formData
        name: NodeID
        required: true
        type: string
      - description: Event topic
        in: formData
        name: topic
        type: string
      - description: Event description
        in: formData
        name: description
        type: string
      - description: Organization responsible for content (e.g. /fac/eng/com)
        in: formData
        name: org_of_content
        type: string
      - description: Event status
        enum:
        - active
        - draft
        - inactive
        in: formData
        name: status
        type: string
      - description: Maximum participants
        in: formData
        name: max_participation
        type: integer
      - description: Whether event has a form (true/false)
        in: formData
        name: have_form
        type: boolean
      - description: Organization path of the posting role
        in: formData
        name: postedAs.org_path
        required: true
        type: string
      - description: Position key of the posting role
        in: formData
        name: postedAs.position_key
        required: true
        type: string
      - description: Visibility JSON string, e.g., {\
        in: formData
        name: visibility
        type: string
      - description: Schedules JSON array, e.g., [{\
        in: formData
        name: schedules
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Event created successfully
          schema:
            $ref: '#/definitions/dto.EventCreateResult'
        "400":
          description: Bad request (invalid input)
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'Forbidden: Cannot post as this role'
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new event
      tags:
      - events
  /event/{event_id}:
    delete:
      description: Mark event as hidden (status=inactive)
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      responses:
        "200":
          description: Event deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Soft delete an event
      tags:
      - events
    get:
      description: Retrieve full details of an event including schedules and form
        ID
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventDetail'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get event detail
      tags:
      - events
    patch:
      consumes:
      - application/json
      - multipart/form-data
      description: Update any field of an event, including image and schedules
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Event fields to update (JSON)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.EventRequestDTO'
      - description: Optional event image file
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Event updated successfully
          schema:
            $ref: '#/definitions/dto.EventRequestDTO'
        "400":
          description: Bad request (invalid event_id or request body)
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'Forbidden: Cannot post as this role'
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an event
      tags:
      - events
...
  /event/participate/{event_id}:
    post:
      description: Join an event directly if it does not require a form
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Participation successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Participate in an event (no form required)
      tags:
      - events
...
swagger: "2.0"